# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_recognize_image_api_v1_recognize_post": {
        "properties": {
          "image": {
            "format": "binary",
            "title": "Image",
            "type": "string"
          },
          "sample_consent": {
            "default": false,
            "title": "Sample Consent",
            "type": "boolean"
          },
          "sample_scene": {
            "default": "other",
            "maxLength": 100,
            "title": "Sample Scene",
            "type": "string"
          },
          "scene": {
            "default": "camera",
            "enum": [
              "camera",
              "album"
            ],
            "title": "Scene",
            "type": "string"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_recognize_image_api_v1_recognize_post",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "code",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecognitionCandidate": {
        "additionalProperties": false,
        "properties": {
          "character_id": {
            "title": "Character Id",
            "type": "string"
          },
          "cn_name": {
            "title": "Cn Name",
            "type": "string"
          },
          "culture_summary": {
            "title": "Culture Summary",
            "type": "string"
          },
          "image_url": {
            "default": "",
            "title": "Image Url",
            "type": "string"
          },
          "provider_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Score"
          },
          "source_ref": {
            "title": "Source Ref",
            "type": "string"
          },
          "variants": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Variants",
            "type": "array"
          }
        },
        "required": [
          "character_id",
          "cn_name",
          "culture_summary",
          "source_ref"
        ],
        "title": "RecognitionCandidate",
        "type": "object"
      },
      "RecognitionResponse": {
        "additionalProperties": false,
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/RecognitionCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "NEED_USER_CONFIRM",
              "UNKNOWN"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "status",
          "provider",
          "model_version",
          "latency_ms",
          "candidates"
        ],
        "title": "RecognitionResponse",
        "type": "object"
      },
...
      "SystemConfigUpdate": {
        "properties": {
          "clear_provider_api_key": {
            "default": false,
            "title": "Clear Provider Api Key",
            "type": "boolean"
          },
          "map_center_latitude": {
            "default": 26.875,
            "maximum": 54.0,
            "minimum": 3.0,
            "title": "Map Center Latitude",
            "type": "number"
          },
          "map_center_longitude": {
            "default": 100.235,
            "maximum": 135.0,
            "minimum": 73.0,
            "title": "Map Center Longitude",
            "type": "number"
          },
          "map_default_zoom": {
            "default": 12,
            "maximum": 18.0,
            "minimum": 3.0,
            "title": "Map Default Zoom",
            "type": "integer"
          },
          "map_web_key": {
            "maxLength": 256,
            "title": "Map Web Key",
            "type": "string"
          },
          "provider_api_key": {
            "default": "",
            "maxLength": 4096,
            "title": "Provider Api Key",
            "type": "string"
          },
          "provider_endpoint": {
            "maxLength": 500,
            "title": "Provider Endpoint",
            "type": "string"
          },
          "provider_model": {
            "maxLength": 160,
            "title": "Provider Model",
            "type": "string"
          },
          "provider_name": {
            "maxLength": 40,
            "title": "Provider Name",
            "type": "string"
          },
          "provider_timeout_seconds": {
            "exclusiveMinimum": 0.0,
            "maximum": 60.0,
            "title": "Provider Timeout Seconds",
            "type": "number"
          }
        },
        "required": [
          "provider_name",
          "provider_endpoint",
          "provider_model",
          "provider_timeout_seconds",
          "map_web_key"
        ],
        "title": "SystemConfigUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Dongba Lijiang API",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/system-config": {
      "get": {
        "operationId": "get_system_config_api_v1_admin_system_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get System Config",
        "tags": [
          "System configuration"
        ]
      },
      "put": {
        "operationId": "save_system_config_api_v1_admin_system_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save System Config",
        "tags": [
          "System configuration"
        ]
      }
    },
...
    "/api/v1/characters/{entity_id}": {
      "get": {
        "operationId": "get_characters_api_v1_characters__entity_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entity_id",
            "required": true,
            "schema": {
              "title": "Entity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Characters",
        "tags": [
          "Business"
        ]
      }
    },
...
    "/api/v1/recognize": {
      "post": {
        "operationId": "recognize_image_api_v1_recognize_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_recognize_image_api_v1_recognize_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecognitionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Gateway Timeout"
          }
        },
        "summary": "Recognize Image"
      }
    },
...
  }
}